Add Try<SuccessType>
This is similar to Result<SuccessType, ErrorType> but tailored to purely
being used to wrap and unwrap either a SuccessType or an exception. This
means we don’t specify the ErrorType, and is thus more generic.

Specifically it can be used in protocols without binding an
implementation to a specific ErrorType or requiring it to be specified
with an associated type.